Create folder contexts and CitiesContext file
