+++ src/contexts/CitiesContext.jsx
@@ -1,0 +1,38 @@
+import { createContext, useEffect, useState } from "react";
+
+const CitiesContext = createContext();
+
+/* eslint-disable react/prop-types */
+
+function CitiesProvider({ children }) {
+  const [cities, setCities] = useState([]);
+  const [isLoading, setIsLoading] = useState(false);
+
+  useEffect(function () {
+    async function fetchCities() {
+      try {
+        setIsLoading(true);
+        const res = await fetch(`http://localhost:8001/cities`);
+        const data = await res.json();
+        setCities(data);
+      } catch {
+        alert("There is an error loading data");
+      } finally {
+        setIsLoading(false);
+      }
+    }
+    fetchCities();
+  }, []);
+  return (
+    <CitiesContext.Provider
+      value={{
+        cities,
+        isLoading,
+      }}
+    >
+      {children}
+    </CitiesContext.Provider>
+  );
+}
+
+export { CitiesProvider };
